회원가입 로직에 있는 평가하기 화면 리스트 UI 작업  - 서버에서 가져온 데이터를 ui로 출력  - AuthInterceptor에서 날짜 체크 이슈 수정 이슈코드 7bbecf4 참고

+++ app/src/main/java/com/leisurely/people/enjoyd/data/remote/data/response/evaluation/DramaEvaluationResponse.kt
@@ -12,7 +12,7 @@ import com.google.gson.annotations.SerializedName
  */
 data class DramaEvaluationResponse(
     @SerializedName("avg_rating")
-    val avgRating: Int,
+    val avgRating: Float,
     @SerializedName("pk")
     val pk: Int,
     @SerializedName("poster")
